youtube-course/mongodb/modules/module-1-quickstart/requirements.txt
# MongoDB RAG Course Requirements
pymongo==4.6.1
openai==1.12.0
voyageai==0.2.3
langchain==0.1.7
langchain-community==0.0.21
python-dotenv==1.0.0
gradio==4.19.1
fastapi==0.109.2
uvicorn==0.27.1
pydantic==2.6.1
